ProvideSupport-Help32-ReferenceNonNumeric
As part of the consensus process, the following sentence was removed from Provide Support (line 828):
"Where an understanding of mathematics is not a primary requirement for using this content, reinforce numbers with non-numerical values." As this SC moves forward, this phrase needs to be reevaluated and discussed with the WCAG group because it specifically addresses the needs of people with dyscalculia.  The original Help issue can be found at the following address:  https://github.com/w3c/wcag21/issues/32.